Keep current due date when changing "Working days only"
That's how it was working before, and it was changed with the work done
on #62891. This has been restored.

Also remove the need to force touching the start date in date picker
when the work package is scheduled automatically: it's not needed
anymore as the start date is always recalculated when in automatic mode.
Plus it was causing issues when switching back and forth between
scheduling mode as the start date would stick instead of being
recalculated.