refactor: name routes after DatePickerController's name
Having the route being `datepicker_dialog_content` leads to looking for
a `DatepickerDialogContentController`, but that would be wrong as the
controller is `DatePickerController`. That's not coherent. That's why
routes were moved to `date_picker` so that the routes and the controller
are named the same

Also moved the `new` action to the collection level so that it does not
need a dummy `work_package_id` with value `new` anymore.