Fix flaky spec
Spec was ./spec/features/work_packages/scheduling/scheduling_mode_spec.rb:139
on job run https://github.com/opf/openproject/actions/runs/14519804438/job/40741624438

In `DateEditField#set_value`, the input fields of the date picker were
cleared and then the values were set. Problem is that when clearining
fields it does not wait for the preview to be applied: when the preview
comes back, it morphs the fields to set empty values but the
`#set_value` method has already started to fill the values, so the start
date field is overwritten with an empty value. Then the test fails
because the start date field is not the one expected.

The solution is to wait for the changes to be applied after clearing the
fields.

I wonder if that may be hiding a real bug where user could have their
input be reset while typing.