Fix flaky spec
Failing spec ./spec/features/activities/work_package/activities_spec.rb:539
on job run https://github.com/opf/openproject/actions/runs/14531919190/job/40773106633

The test did not correctly wait for the second journals changes to be
displayed. It was waiting with
`activity_tab.expect_journal_changed_attribute(text: "Subject")` but as
the subject was already changed in the first journal entry, it was
returning immediately and did not wait for the second journal entry to
be updated.

Waiting for `activity_tab.expect_journal_changed_attribute(text: "A new
subject")` instead forces the test to wait for the second journal entry
to be displayed and fixed the flakyness.