Fix flaky specs
Failing spec is ./spec/features/types/form_configuration_query_spec.rb:174
on job run https://github.com/opf/openproject/actions/runs/14513528848/job/40717420434

The spec seems flaky because the form configuration modal is opened
while the page is reloaded, which closes the modal and then the elements
inside the modal cannot be clicked anymore.

This is fixed by calling `expect_and_dismiss_flash` instead of
`expect_flash` in the specs so that it waits for the save to complete.
If not dismissed, it does not wait the second time, which is the case in
the flaky spec.

I also changed `add_query_group` to yield the modal to the caller so
that the caller can interact with the modal before saving it. It's
better than creating it, closing it, then reopeining it again with
`edit_query_group` to make modifications. I thought it was the culprit
for the flakiness but actually no. But the code read better so I kept
it.