Fix flaky spec
Failing spec is spec/features/work_packages/custom_actions/custom_actions_spec.rb:140
on job run https://github.com/opf/openproject/actions/runs/14613189364/job/40995333431?pr=18684

It occurs because the custom actions stored in the cache are overwritten
with some data after fetching the work package: when the work pacakge is
fetched, it has a `links` property with multiple things, including a
`customActions` array. This array of custom actions only has a name
property, so when the cache is updated, the custom actions are
overwritten with only the name property.

As there is only the name property, the description is lost and the
title of the button is set to the name instead of being set to the
description.

Sometimes it works because the work package is fetched before the
action are fetched, and sometimes it doesn't.

The fix is to keep custom actions in the cache if they exist. This
probably introduces bugs because if the custom actions array is very
detailed, we want to override it, but that's a try.

So the fix applied here is absolutely not clean and should be revisited.
It's a basis for a discussion about how we want to handle this.

The title has been moved to a proeprty so that we can stop the debugger
on it.