Fix flaky unit test
Failing spec is ./spec/controllers/work_package_relations_controller_spec.rb:133
on job run https://github.com/opf/openproject/actions/runs/14646323170/job/41101314595

Reproduction command:

    rspec './spec/controllers/work_package_relations_controller_spec.rb[1:4:1,1:5:2]' --seed 37278

The `relation` instance is destroyed in a test, and reloaded for the
next test as we are using `shared_let`. But when reloading a destroyed
record, it remains destroyed. When updating it, it tries to save it
again, including the related work packages but they already exist in
database. This leads to the error "ActiveRecord::RecordInvalid:
Validation failed: Related work package has already been taken.".

To fix this, we use `refind: true` to reload the relation using a
database `find`.