Fix flaky spec
spec is ./modules/reporting/spec/features/update_entries_spec.rb:72
job run is https://github.com/opf/openproject/actions/runs/14646323170/job/41135997810

It was failing when submitting the time logging modal before the work
package was loaded. As it was not loaded, it was empty, and as it was
empty, it was displaying an error "Work package is invalid" and the time
was not updated.

As the time was not updated, the spec was failing looking for "2.00
hours" (the new value) but only getting "5.00 hours" (the old value).

Fix is to wait for the autocompleter to be fully loaded before
proceeding with submit. This is done by waiting for `ng-spinner-loader`
to not be visible anymore.