Use certificate fingerprint as a fallback only
As far as I can tell from existing code and forms, we only want to use
the fingerprint as a fallback and not as a primary means of authentication.

However, the HashBuilder would have only returned the fingerprint, if both
a fingerprint and a certificate were present, which sadly always seems to be
the case, because even when creating a new provider, we fill the fingerprint
from the provided certificate.

Since the omniauth configuration happens from the provider hash, this means
that we so far only ever used the fingerprint for validation.