Correctly handle users unable to authenticate at a storage
The previous code assumed that if the authentication method selector
doesn't choose SSO, it must have chosen OAuth 2.0, which results in
rendering an ensure connection link. However, it's also possible that
the method selector chooses NO authentication method, because the
user simply can't authenticate at the storage. In this case rendering
an ensure connection link would fail.

This was not detected earlier due to missing tests.