Fix parsing of multiple certificates from metadata
Once the XML contains multiple certificates, the metadata hash
has a nil idp_cert. The idp_cert_multi key is always populated. However,
we never tried parsing data from there, so it was lost.

The way that the Saml::Provider right now represents multiple certificates
is to concatenate them inside the idp_cert, so the metadata parsing
has been adapted to use that.