[60580] Fix error 500 when bulk editing work packages
https://community.openproject.org/wp/60580

When multiple CKEditor instances are on the same form, each of them
intercepts the submit event and submit the form, leading to multiple
identical parallel submit requests.

This happens for instance on the bulk edit page when the work packages
have long text custom fields because this adds a ckeditor in the form.
The form is submitted multiple times and only one request will succeed,
the other will produce a 500. That's why a 500 error is sometimes
visible on bulk edit.

The fix is to track the "in-flight" state per form in a variable shared
with all instances of the component instead of storing it in an instance
variable.