[63831] Fix error 500 when bulk deleting related work packages
https://community.openproject.org/wp/63831

When bulk deleting work packages, successors of descendants are
rescheduled, because as they have one less predecessor, they may need to
be switched to manual scheduling mode.

Problem: if one of these successor to reschedule is also a child of the
deleted work package, then it cannot be rescheduled: it's deleted. When
trying to rescheduled and save it, the error
`ActiveRecord::StaleObjectError` is raised.

Solution: only reschedule successors that are not descendants of the
deleted work package.