Make feature tests more stable
Following advices on https://github.com/teamcapybara/capybara/issues/2796

This occurs sometimes that some tests are failing in headless mode: the
DOM is not correctly reflecting the state of the page until a screenshot
is taken.

This happened in ./spec/features/work_packages/table/queries/query_menu_refresh_spec.rb:44
in job https://github.com/opf/openproject/actions/runs/14965572620/job/42034828494

The screenshot shows that the query menu item is here, but before the
screenshot is taken, capybara cannot find it.

This is a shot in the dark. I have no idea if that will really fix the issue mentioned above.