Fix flaky spec
spec is modules/meeting/spec/features/meeting_backlogs_spec.rb:128
failing job is https://github.com/opf/openproject/actions/runs/14969474602/job/42047670034

Problem is that the menu can disappear before the action is clicked. It
disappears when the message "The meeting has been updated. Reload?" is
displayed. Sometimes it is just not displayed and I could not figure out
why. The action seems to be swallowed.

Fix is to retry.

It's not perfect yet. It can still fail at
`modules/meeting/spec/features/meeting_backlogs_spec.rb:201` when
calling `clear_backlog` to close the modal. The modal is not there.

I'm not sure how to fix this one.