[#63728] Show error flash on Project update failure
This is a temporary solution while we consider how we surface errors to
users generally. Ideally, since we are already displaying inline
validations - and to avoid unnecessary noise in the UI, we would only
display flash messages when there are fields/attributes with errors that
are NOT visible to the user on this specific Project
Settings>Information form, BUT that the user does have permissions to
edit (albeit elsewhere in the application).