[62847] Remove project filter from hierarchy query link
https://community.openproject.org/wp/62847

Remove the project from the hierarchy link path to make the query
consider work packages from all visible projects.

Before, when clicking on the link to the detailed view of a work
package, the link was using the project work packages path which
incidentally adds a project filter. This prevented children which
belonged to a different project to be displayed (except if they were
belonging to a subproject).