[63475] Preserve unset dates when moving work packages in calendar
https://community.openproject.org/wp/63475

If it's resized, both start and due dates are set.

If it's dragged and dropped and it's a milestone, it's the same as
resizing: both dates are set.

If it's dragged and dropped and it's not a milestone, if it has a
duration, set duration and start date to deal with non-working days, and
if it has no duration, set start date if it was already set, or set the
due date.