[63499] Fix possible SystemStackError
https://community.openproject.org/wp/63499

When there is a cycle in the hierarchy, for instance when drag and
dropping a parent work package to its child from work packages list, or
by bulk editing work packages and setting one of them as the parent, the
scheduling computation would enter an infinite loop (when computing
dates of parent from its children).

This is now fixed by rejecting children that have already been processed.