Allow to pass a scope during token exchange
This is the most basic/immediate implementation of this.
It's now possible to specify scopes that will be requested during
token exchange for storage SSO.

An immediate next step would probably be to rephrase the former
"audience" UI to be more inclusively phrased with regards to containing
settings for token exchange in general.

In the longer run we should also rethink the "backend" code of token exchange:

* The scope is so far not considered at all when figuring out whether we already
  have a matching token, we only compare the audience for that
* We should probably not read audiences of the JWT anymore and only decide between
  performing token exchange (requiring an audience) or not doing it. The currently supported
  use case of configuring an audience that's then expected in the login token does not really
  add any value.
* it's again a question how many authentication details should be stored on a storage
  instead of being stored somewhere away from the storage