remove query behaviour of clearing columns if they match default ones
This behaviour leads to surprising side effects. E.g. by removing a column
the default columns might be matched suddenly so that then the columns change
if the default ones are changed afterwards. Additionally, it becomes impossible
to remove the project column from the global work package list.

The intend of relying on the default columns is still working for newly created queries as well as
for default column (e.g. "all open").