[62426] Preserve scheduling mode when bulk copying work packages
https://community.openproject.org/wp/62426

Copied successors and parent being originally manually scheduled would
automatically be switched to automatic scheduling mode as soon as they
receive their first child or fist predecessor, because that's how it
works. To prevent that, there is a flag `bulk_copy_in_progress: true` in
service state which prevents switching to automatic mode in
`UpdateAncestorsService` if present.

Also copied successors and parent being originally automatically
scheduled would be reset to manual in `SetScheduleService` for parent
without children, or for successor without predecessor, and would never
be switched back to automatic due to `bulk_copy_in_progress: true`. To
prevent that, the automatic mode is forced in `CopyService` when calling
`SetScheduleService`.