Prohibit scheduling automatically in API if no predecessors or children
Only when creating a work package. The only case where it's possible is
when creating a work package with a parent explicitly set, and that
parent is automatically scheduled and has at least one predecessor.

When copying work packages, this restriction is removed as predecessors
and children are added later in the process.