[#62982] consider inactive phases in case of multiple projects
a phase definition can be used in multiple phases. Some of which might
be active and some of which might be inactive - depending on which
projects you show in your query. This commit adds support for this case.

However, there in that case the permissions might have to be checked
differently -> this will be done in a follow-up commit.