Do not require prometheus by default
The code enabling it in config.ru is already requiring
this file, indicating that it was intended to only be enabled
when needed. However, due to automatically requiring gems through
bundler, this gem was still loaded.

We could validate in heap dumps that before this change, the gem
was active and consuming a non-neglectible amount of memory (probably
to do its job) even if it was not enabled. After the change it's not.