[64711] Look up already seeded colors by hexcode in seeders
https://community.openproject.org/wp/64711

When seeding a second time in a different language, it was failing with
the error "ArgumentError: Nothing registered with reference
:default_color_blue (ArgumentError)".

Here is why: the type seeder, as some types are already in the database,
will try to look up existing types by their name attribute to store
their reference in the seeder registry. To get the name attribute, it
computes all types attributes, so it also computes the `color_id`
attribute.

To compute the `color_id` attribute, it will try to look up the color by
its reference name in the seeder registry. As it's not there, it fails.

So now, why is it not there? It's because the colors are already seeded,
so it tries to look up existing ones to store them in the registry. To
find them, it looks them up by their name attribute. But that name is
different if current language is not the one used when the initial
seeding was done. As the lookup returns nothing, it does not store the
color in the registry. That's why looking the color up later fails.

The fix is to look up the colors by their hexcode instead of their name.

And if the colors hexcode has been changed, or if the colors were
deleted, it would fail too. The fix in this case is to catch the lookup
error and ignore it. It's not seeding, it's only looking up existing
records so it's safe to ignore the error.