[64603] Fix milestone handling when switching from manual to automatic
https://community.openproject.org/wp/64603

When it's a milestone being updated, the due date was copied from the
start date. This works when the milestone is in manual scheduling mode,
but when it's in automatic scheduling mode, or when it's switched to
automatic scheduling mode, this copy gives a nil value to the due date,
and due date cannot be set because it's readonly in automatic
scheduling. This lead to the 422 error.

Now the copy of due date from start date is only done if the start date
parameter is present too.