[64636] Prevent importing same token twice
https://community.openproject.org/wp/64636

It's not very strict right now: you can still import the same token
twice if you add some garbage before and/or after, but it strips
whitespaces and for most customer scenarios, it should be enough.

Also even if the same token is added twice, it's not a big deal. It will
still work as expected.