Fix flaky spec
spec was: ./spec/features/work_packages/custom_actions/custom_actions_spec.rb:487
job run was: https://github.com/opf/openproject/actions/runs/15728287105/job/44323158079

The ajax was disabled *before* the page was fully loaded, so the page
was sometimes blank because it was not finished loading.

Fixed by waiting for the custom action button to be displayed on the
page before disabling ajax requests.