Fix flaky spec
failing spec was: ./spec/features/projects/project_custom_fields/overview_page/dialog/attribute_help_texts_spec.rb:112

The database ids are present in the expected text. If that's the first
test to run that's ok, but if not and some attachments were created in
other tests, then this test is doomed to fail.

Fixed by using the real attachment filenames instead of hardcoded
filenames.