Handle non-turbo responses in async-dialog controller
We make a web request, expecting a proper turbo response, but
so far we didn't check that we receive a turbo response. As a result
any non-turbo error page (NGINX error, rails error page, etc.) would
blindly be appended to the page.

By checking the content type, we can at least make sure that the web
server intended for us to handle the response in the way we do.