[64973] Fix error 500 when making a predecessor with children a child of its successor
https://community.openproject.org/wp/64973

The error was caused by a transient circular dependency when computing
futures dates of parent. The circular dependency is transient because it
tries to set the parent, and then an error is displayed saying it's
impossible. Meanwhile the SetAttributesService still tries to compute
the correct dates of the parent, using data with a circular dependency
in it.

This is fixed by marking visited work packages and skipping them when
they have already been visited.