Protect against possible race conditions `CkEditorAugmentedTextareaComponent.saveForm`
This bug is not easily reproducable, but it seems to happen when a user leaves a tab open for a long time, comes back to it and continues to submit a comment- at this point, in one case there were reports of 2-3 parallel form submissions.

The idea here is to prevent any possible race conditions by mitigating possible stale `inFlight` state resulting from angular component lifecycle effects; dismounting and initializing (ngOnDestroy,ngOnInit).

_Expands on e8ec49a34cd6b1603f45d5f700d5bec61cc14e02, Follows #19185_

https://community.openproject.org/wp/63556