[65062] Fix infinite loop when getting automatically scheduled ancestors
https://community.openproject.org/wp/65062

When there is a cycle in the hierarchy, like trying to set the root
parent as a child of its own child, the scheduling dependency would loop
when trying to find all the automatically scheduled ancestors of the
work package.

This is now fixed by marking the visited work packages.