[64203] Fix unclear error message on parent self-assignment
https://community.openproject.org/wp/64203

When updating a work package to be parent of itself, `closure_tree` gem
has hooks to update the work_package_hierarchies table and can detect
cycles. When a cycle is detected, it adds a "You cannot add an ancestor
as a descendant" error on `:parent_id`.

The fix is to remove this error before providing our own error message.