[#64347] fix custom field Primer bugs
In the admin view, when you edit a custom fields name to be empty and
try to save that, the display would incorrectly assume the empty title
from the model, even if the model would be invalid with these changes.
Updated to fetch the name attribute from the database instead.

The bug only happens when you use a Primer component. So for now, this
is only true for Hierarchy and Calculated Value CFs.
Angular prevents this bug from happening since there is no page reload.