Use cautious shell expansions in IMAP job
There are various boolean coercions performed on the values passed to
the IMAP sync job, all of which will evaluate to true when called in the
manner that the docker 'cron' script goes about invoking the job, as
they will be empty strings.

This causes a variety of unexpected behavior, for instance:

- If IMAP_HOST is not set, the job will try to connect to 0.0.0.0 and
complain about it
  - This is doubly frustrating for users of the docker-compose
  distribution, because the .env file does not apply to container
  environments, rather only to variable interpolation in the compose
  file (and maybe build environments).
- Users not intending to leverage the 'move on success' and 'move on
failure' features will experience highly rate-limited email processing
as a result of the sync task attempting to move each processed message
to a folder named '', which is not a valid operation. This has the
effect of causing only one message to be processed per invocation.
- I am sure there are more. The various options which appear to relate
to processing unsolicited email might have unintended side effects if
set to empty strings.