Specify non-abbreviated month format for `OpPrimer::RelativeTimeComponent`
In French, abbreviated month names (like "juil." for "juillet") officially require a period at the end, according to typographic and localization standards. For example, "juil." is the correct abbreviation for "juillet" in French, and the period is part of the abbreviation.

As out translation file ends with a period, the date component would add a seemingly redundant period at the end which looks like a double period ("..")

> ..notification for this work package le 26 juil..