Fix broken `rake db:seed`
It was failing with the error "cannot generate demo seed data without
setting locale first".

When seeding, there is a mechanism in place to ensure the seed data is
not accessed before the language has been set. The new mechanism
introduced in b0716517e47291de672 to check if all references are
available in `applicable?` accesses the seed data before the language
has been set for `RootSeeder`, making the `rake db:seed` task fail.

The fix is to bypass the `applicable?` check for `RootSeeder` and return
`true`.

Also it was not detected by our test suite because `#seed_data!` was
used instead of `#seed!`. This is now fixed.