Add enterprise checks for SCIM API
Including a banner for the UI (currently using a placeholder image)
and a check for SCIM API calls.

While it's theoretically possible to setup a client without an enterprise
token (new/edit/create/update are not guarded), there is no click path to
create a client and eventually the API requests would be blocked.