Show error message when errors occur
While opening the project storage we try to create the folder
if it doesn't exist yet and we ensure permissions to that folder.
However, previously we wouldn't have recognized that an error happened,
because the result was not indicating to be a failure.
This has now been fixed.

The add_error method on the services will now make sure to fail the result,
which means we can't forget to set the result to failed. To keep the behaviour
mostly compatible with the previous code, the ManagedFolderSyncService will now
ignore errors during folder creation and try to set permissions regardless.
This was previously effectively happening, because most error conditions didn't
fail the result anyways.