[63191] Show % Complete sum in more cases
https://community.openproject.org/wp/63191

Before, % Complete sum was calculated based on work and remaining work
sums, meaning work packages with only % Complete is set are ignored.

Now, if "Calculation of % Complete hierarchy totals" is set to "Simple
average", the % Complete sum is calculated as an average of the %
Complete values of work packages having % Complete set.

If "Calculation of % Complete hierarchy totals" is set to "Weighted by
work", the calculation is not changed: it is still based on work and
remaining work sums.