bug/63545 When editing a comment in the activity tab, the form can be force closed
The activities tab polling mechanism updates newly updated journals, based on the last update timestamp. In case an attachment is added, or the journal is updated in any other places- the UI item would be updated\ on the next poll, overriding the edit form.

This commit adds an exclusion for any forms that are in "edit" state- it does not yet take into account conflict resolution.

https://community.openproject.org/wp/63545
https://community.openproject.org/wp/65680