Fix flaky spec
failing spec: ./spec/features/work_packages/details/relations/hierarchy_spec.rb:151
failing job: https://github.com/opf/openproject/actions/runs/16163863659/job/45620997168?pr=19479

The data-update-work-package attribute is read to know if the relations
tab and its counter should be reloaded. Sometimes the event target is
the html document instead of the form element, probably due to some
other reloading of the relations tab (it reloads a lot actually).

The fix is to read the form element from the formSubmission property
given by the turbo:submit-end event. It's guaranteed to be the form
element from which the event was triggered.