WIP: Add rudimentary service for requests towards Haystack
The general idea is that we ask Haystack to create a report
for us and provide it with an OAuth access token that allows making
API requests in the name of the current user for a limited amount of
time.

Code provided here is intended to be a proof-of-concept that was created
during a one-week hackathon.