Store Pmflex hints in database and pull from API
Instead of hardcoding certain checks into our codebase,
we will now call an API to get the hints and store them locally,
so that the API does not have to be called all the time.

On the other side of the API, an LLM will probably come up with
all the answers.