Avoid resetting the scroll position when updating the relations tab
This was causing a "scroll flickering": it scrolled to the added item as
the response to the creation of the relation returned an element having
the `data-work-packages--relations-tab--scroll-target="scrollToRow"`
attribute, then immediately scrolled back to the top as the
relations_tab content was requested again via turbo stream and replaced
the relation tab.

By using `morph` during the replacement, the scroll position is preserved.