allow external PRs to trigger SaaS tests
changing the event to pull_request_target checks out the branch's (i.e. release or dev) code instead of the PR's code. This is fine in this case because we don't actually use any of the code. We just trigger another workflow. As for security, this is also fine. The reason this is not allowed for external PRs, and why we need to use pull_request_target rather than pull_request, is that there is a risk that the external PR's code wants to read and leak secrets. Using pull_request_target prevents that. Now the downstream workflow will still check out this external PR's branch. But it won't make use of any of the workflows. It's actually testing the code. So there is no risk here.