Fix features reliant on Capybara hidden heuristics
Capybara, when using #text or similar methods, typically only returns
text that is considered visible on the page. However, the definition of
"visible" depends on both the underlying driver and Capybara's
heuristics.

As such, Capybara + Cuprite consider the following rule in our
`.hidden-for-sighted` implementation's rule as "invisible":

    left: -10000px

However the Primer `.sr-only` implementation is considered "visible":

    clip-path: rect(0 0 0 0)

This commit updates the affected feature specs to account for this
difference in visibility detection. The tests now properly handle
screen reader only text that Capybara considers "visible" when using
Primer's `.sr-only` class instead of the custom `.hidden-for-sighted`
implementation.