[#64348] Remove TODO comment
To implement this properly, we would have to either cache results of
previous validation checks, or we would have to perform certain queries
and operations again. For example, we would have to check which of the
surplus CFs do not have an allowed field format and add an error for
them.

After that, we would have to find all fields causing a circular
reference and add an error for them.

Same for fields not visible to the user, and so on and so on.

What I am trying to get at is: This would complicate matters quite a
lot. However, if all goes well, users will never add these fields to a
formula, since it is never suggested in the first place. So while being
nice and helpful, it hopefully will not happen very often. Therefore, I
would keep it vague for the first iteration and observe what the users
do with this feature.