Disable pattern input instead of hiding it for enterprise
This required to allow disabling the pattern input. For that
it received the additional FormControl-input class, which makes
sense, because it behaves like a form input and made it react
to the disabled attribute.

A side effect of the new class seems to be that some paddings changed
slightly, but it still looks good to me.