Show storage configuration in health status download
Limiting that to non-confidential configuration, so that
people do not leak passwords or similar to us, when they use
the report to request support.
On the other hand, we include as many other details as possible,
to get as much of an idea on how the storage could be misconfigured
as possible.