Clarify APIv3 docs
When docs started directly in the tables of properties, without any introductory
words, the linked tables appeared directly under the headline "ENTITY NAMES" (e.g. "Users").
This gave the wrong impression, that the tables would explain the links and properties of
the collection endpoint of those classes.

This was fixed by adding a clarifying introductory paragraph to endpoint docs that
didn't have an intro paragraph already.