Show proper names of tokens disabled for current type
This is for example the case, when a custom field
is not shown for a specific type. In this case, we have
all the information available to show its proper name,
we just don't want to suggest to the user to use it and
we don't want to indicate it as a valid choice.