Fix flaky spec
failing spec: `spec/features/work_packages/details/date_editor_spec.rb:481`
failing job: https://github.com/opf/openproject/actions/runs/17148953008/job/48650666950?pr=20026

The spec does not wait long enough for the page to load. It tries to
activate the combinedDate field, and this one has internal checks about
its state to ensure it has really been clicked. So it's clicked, but the
loading takes some time, so it tries again. But when it tries again, the
display field has already been masked, so it can't be clicked a second
time, and the spec fails. (See `#activate!` in
`spec/support/edit_fields/edit_field.rb` for more details.)

When it's fast enough, the test passes, hence the flickering.

To fix it, we wait for the page to be fully loaded before clicking the
date field. This way the activation of the field and the loading of the
date picker is hopefully done faster.

To wait for loading, we wait for an element inside the main work package
page component, and wait for the loader placeholders to go away in the
activity tab.